Move the edit token styles into their own file
Now that the modal is gone, the styles have some unused code. This makes it possible to find & delete it.
